sync/requirements.txt
requests~=2.31.0
pytz~=2021.3
Flask~=2.0.2
json2html~=1.3.0
Werkzeug~=2.0.2
numpy~=1.22.1
networkx~=2.6.3
gunicorn==20.1.0
pycryptodome==3.14.1
psycopg2-binary==2.9.6
scipy~=1.10.1
prometheus-client~=0.19.0